issue-164: Separating table configuration for better code readble
